requirements.txt
Pillow>=10.3.0,<10.4
django>=5.0.2,<5.1
djangorestframework>=3.15.0,<3.16
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
django-ckeditor-5>=0.2.13,<0.3
django-cors-headers>=4.3.1,<4.4
django-jazzmin>=3.0.0,<3.1
django-jsonform>=2.22.0,<2.23
uwsgi>=2.0.20,<2.1
PyJWT>=2.8.0,<2.9
requests>=2.32.3,<2.33
oauthlib>=3.2.2,<3.3
django-mail-admin>=0.3.2,<0.4
itsdangerous>=2.2.0,<2.3
django-appmail>=6.0,<6.1